[refactor] SavedBook을 Book 패키지로 이동 -> 저장 관련 로직 포함 (#156)
