+++ src/test/java/konkuk/thip/book/adapter/in/web/BookChangeSavedControllerTest.java
@@ -7,8 +7,8 @@ import konkuk.thip.book.adapter.out.persistence.repository.BookJpaRepository;
 import konkuk.thip.common.exception.code.ErrorCode;
 import konkuk.thip.common.security.util.JwtUtil;
 import konkuk.thip.common.util.TestEntityFactory;
-import konkuk.thip.saved.adapter.out.jpa.SavedBookJpaEntity;
-import konkuk.thip.saved.adapter.out.persistence.repository.SavedBookJpaRepository;
+import konkuk.thip.book.adapter.out.jpa.SavedBookJpaEntity;
+import konkuk.thip.book.adapter.out.persistence.repository.SavedBookJpaRepository;
 import konkuk.thip.user.adapter.out.jpa.AliasJpaEntity;
 import konkuk.thip.user.adapter.out.jpa.UserJpaEntity;
 import konkuk.thip.user.adapter.out.jpa.UserRole;
