+++ src/test/java/konkuk/thip/book/adapter/in/web/BookDetailSearchControllerTest.java
@@ -13,8 +13,8 @@ import konkuk.thip.room.adapter.out.jpa.RoomParticipantRole;
 import konkuk.thip.room.adapter.out.persistence.repository.RoomJpaRepository;
 import konkuk.thip.room.adapter.out.persistence.repository.category.CategoryJpaRepository;
 import konkuk.thip.room.adapter.out.persistence.repository.roomparticipant.RoomParticipantJpaRepository;
-import konkuk.thip.saved.adapter.out.jpa.SavedBookJpaEntity;
-import konkuk.thip.saved.adapter.out.persistence.repository.SavedBookJpaRepository;
+import konkuk.thip.book.adapter.out.jpa.SavedBookJpaEntity;
+import konkuk.thip.book.adapter.out.persistence.repository.SavedBookJpaRepository;
 import konkuk.thip.user.adapter.out.jpa.AliasJpaEntity;
 import konkuk.thip.user.adapter.out.jpa.UserJpaEntity;
 import konkuk.thip.user.adapter.out.jpa.UserRole;
